Shrink a couple bytes of struct attr

+++ bgpd/bgp_attr.h
@@ -209,7 +209,7 @@ struct attr {
 
 	/* has the route-map changed any attribute?
 	   Used on the peer outbound side. */
-	uint32_t rmap_change_flags;
+	uint16_t rmap_change_flags;
 
 	/* Multi-Protocol Nexthop, AFI IPv6 */
 	struct in6_addr mp_nexthop_global;
@@ -293,7 +293,7 @@ struct attr {
 	/* EVPN local router-mac */
 	struct ethaddr rmac;
 
-	uint16_t encap_tunneltype;
+	uint8_t encap_tunneltype;
 
 	/* rmap set table */
 	uint32_t rmap_table_id;
